Extract swarm
